pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdf-cot-monitorability"
version = "0.1.0"
description = "SDF CoT Monitorability"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "inspect-ai>=0.3.50",
    "openai>=1.58.0",
    "anthropic>=0.42.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "rich>=13.0.0",
    "fire>=0.7.0",
    "pydantic>=2.10.0",
    "aiofiles>=22.0",
    "impossiblebench",
    "false-facts",
]

[dependency-groups]
dev = [
    "poethepoet>=0.38.0",
    "pyright>=1.1.407",
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sdf_cot_monitorability*"]
exclude = ["external*", "results*", "configs*", "scripts*", "tests*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
# a sensible starter set; adjust as you like
select = ["A","E","F","I","N","UP","B","T100"]

[tool.pytest.ini_options]
pythonpath = [".", "sdf_cot_monitorability"]
addopts = "-q"

[tool.poe.tasks]
fmt = "ruff format"
lint = "ruff check --fix"
check = "pyright"
test = "pytest"
all = [{ref="fmt"}, {ref="lint"}, {ref="check"}, {ref="test"}]

[tool.uv]
# Override conflicting dependencies between false-facts and impossiblebench
# false-facts has many pinned versions that are too old
override-dependencies = [
    "datasets>=4.0.0",      # false-facts wants 3.2.0, impossiblebench wants 4.0+
    "tiktoken>=0.11.0",     # false-facts wants 0.8.0, impossiblebench wants 0.11+
    "aiofiles>=22.0,<24.0", # false-facts/gradio wants <24.0
    "pydantic>=2.11.4",     # false-facts wants 2.10.4, inspect-ai wants 2.11.4+
    "bitsandbytes ; sys_platform != 'darwin'",  # Skip on macOS (no ARM wheels)
]

[tool.uv.sources]
impossiblebench = { path = "external/impossiblebench", editable = true }
false-facts = { path = "external/false-facts", editable = true }
